move all atrules in the proper directory structure
