+++ org/w3c/css/atrules/css/AtRuleNamespace.java
@@ -7,7 +7,9 @@
  * AtRuleMedia.java
  * $Id$
  */
-package org.w3c.css.parser;
+package org.w3c.css.atrules.css;
+
+import org.w3c.css.parser.AtRule;
 
 /**
  * This class manages all imports
