+++ org/w3c/css/atrules/css/AtRulePage.java
@@ -8,8 +8,9 @@
  * AtRulePage.java
  * $Id$
  */
-package org.w3c.css.parser;
+package org.w3c.css.atrules.css;
 
+import org.w3c.css.parser.AtRule;
 import org.w3c.css.util.ApplContext;
 import org.w3c.css.util.InvalidParamException;
