+++ org/w3c/css/atrules/css/AtRulePhoneticAlphabet.java
@@ -8,8 +8,9 @@
  * AtRulePhoneticAlphabet.java
  * $Id$
  */
-package org.w3c.css.parser;
+package org.w3c.css.atrules.css;
 
+import org.w3c.css.parser.AtRule;
 import org.w3c.css.util.ApplContext;
 
 /**
