+++ org/w3c/css/atrules/css/AtRulePreference.java
@@ -5,19 +5,21 @@
 // (c) COPYRIGHT 1995-2000  World Wide Web Consortium (MIT, INRIA, Keio University)
 // Please first read the full copyright statement at
 // http://www.w3.org/Consortium/Legal/copyright-software-19980720
-package org.w3c.css.properties.svg;
+
+package org.w3c.css.atrules.css;
 
 import org.w3c.css.parser.AtRule;
 
-public class AtRuleColorProfile extends AtRule {
+@Deprecated
+public class AtRulePreference extends AtRule {
 
     static int internal = 0;
     int hash;
 
     /**
-     * Create a new AtRuleColorProfile
+     * Create a new AtRulePreference
      */
-    public AtRuleColorProfile() {
+    public AtRulePreference() {
         hash = ++internal;
     }
 
@@ -25,28 +27,28 @@ public class AtRuleColorProfile extends AtRule {
      * Returns the at rule keyword
      */
     public String keyword() {
-        return "color-profile";
+        return "preference";
     }
 
     /**
      * The second must be exactly the same as this one
      */
     public boolean canApply(AtRule atRule) {
-        return (atRule instanceof AtRuleColorProfile);
+        return (atRule instanceof AtRuleFontFace);
     }
 
     /**
-     * Return true if other is an instance of AtRuleColorProfile
+     * Return true if other is an instance of AtRulePreference
      */
     public boolean equals(Object other) {
-        return (other instanceof AtRuleColorProfile);
+        return (other instanceof AtRulePreference);
     }
 
     /**
      * The second must only match this one
      */
     public boolean canMatch(AtRule atRule) {
-        return (atRule instanceof AtRuleColorProfile);
+        return (atRule instanceof AtRulePreference);
     }
 
     /**
