+++ org/w3c/css/parser/analyzer/CssParser.jj
@@ -64,15 +64,15 @@ import org.w3c.css.parser.CssErrorToken;
 import org.w3c.css.parser.CssSelectors;
 import org.w3c.css.parser.CssParseException;
 import org.w3c.css.parser.AtRule;
-import org.w3c.css.parser.AtRuleCounterStyle;
+import org.w3c.css.atrules.css.AtRuleCounterStyle;
 import org.w3c.css.atrules.css.AtRuleMedia;
 import org.w3c.css.atrules.css.media.MediaFeature;
-import org.w3c.css.parser.AtRuleFontFace;
-import org.w3c.css.parser.AtRuleKeyframes;
-import org.w3c.css.parser.AtRulePage;
-import org.w3c.css.parser.AtRulePreference;
-import org.w3c.css.parser.AtRulePhoneticAlphabet;
-import org.w3c.css.properties.svg.AtRuleColorProfile;
+import org.w3c.css.atrules.css.AtRuleFontFace;
+import org.w3c.css.atrules.css.AtRuleKeyframes;
+import org.w3c.css.atrules.css.AtRulePage;
+import org.w3c.css.atrules.css.AtRulePreference;
+import org.w3c.css.atrules.css.AtRulePhoneticAlphabet;
+import org.w3c.css.atrules.svg.AtRuleColorProfile;
 import org.w3c.css.util.ApplContext;
 import org.w3c.css.util.CssProfile;
 import org.w3c.css.util.CssVersion;
